end auth

+++ Repositories/Abstract/IRepos.cs
@@ -4,6 +4,6 @@ public interface IRepos<T> where T : Abstract.IModel
 {
     public Task<T?> Add(T model);
     public Task<T?> Update(T model);
-    public Task<T?> GetModel(int tgID);
+    public Task<T?> GetModel(long tgID);
     public Task<List<T>?> GetAllModel();
 }
