+++ Repositories/Wallet/WalletRepos.cs
@@ -13,7 +13,7 @@ public class WalletRepos : IRepos<WalletModel>
     {
         this.tgBotOrderContext = tgBotOrderContext;
     }
-    public async Task<WalletModel?> CreateModel(int tgID, string address, float balance)
+    public async Task<WalletModel?> CreateModel(long tgID, string address, float balance)
     {
         WalletModel walletModel = new WalletModel(tgID, address, balance);
         return walletModel;
@@ -23,7 +23,7 @@ public class WalletRepos : IRepos<WalletModel>
         List<WalletModel> walletModels = new List<WalletModel>();
         return walletModels;
     }
-    public async Task<WalletModel?> GetModel(int tgID)
+    public async Task<WalletModel?> GetModel(long tgID)
     {
         WalletModel walletModel = new WalletModel(tgID, "", 0);
         return walletModel;
